Moved modelling to new folder within `app/components`. Introduced PSF convolution from PSF contained within subject metadata.
